Adjust python dependency for sage ebuilds.
Now we have two depend atoms for python. One is ~2.6.5:2.6 which ensures
that the 2.6 slot is filled by a 2.6.5 ebuild and not a later version.  The
other is >=2.6.5-r99:2.6 which ensures that this 2.6.5 version is at least
revision 99, so it should come from this overlay and contain the required
fix.