Employ USE flag instead of revision numbers to ensure python fix.
Expressing the presence of a given feature (here the fix for issue7689) in
terms or an offset to revision numbers proves to be tricky to maintain.
Using USE flags to express such dependencies has proven rather robust for
kde-sunset.  The ebuilds from the overlay simply offer an use flag which the
main portage tree doesn't provide, and the sage ebuilds can depend on it.