Import maxima 5.30.0 from main tree with modifications.
The part building a single fasl file for ecl has been re-enabled bur worded
so it works with the modifications from ecl 2.26.71 ("." no longer in source
registry by default) as well as changes to the way the user cache is
handled.  Also include a user-contributed patch for the maxima texinfo
manual, to make it work with texinfo 5.1.

Hopefully this fixes sage-on-gentoo issue cschwan#226
as well as gentoo bug 466730.